requirements.txt
requests~=2.23.0
module~=0.0.4
prettytable~=2.2.1
shodan~=1.27.0
zoomeye~=2.0.4.6.13
openpyxl~=3.0.7
PyMySQL~=0.9.2
PyYAML~=5.4.1
mmh3~=3.0.0
urllib3~=1.25.3
bs4~=0.0.1
beautifulsoup4~=4.11.1
XlsxWriter~=1.4.4
ipaddress~=1.0.23
colorama~=0.4.4